pyproject.toml
[project]
name = "streamlit-kafka"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "confluent-kafka>=2.10.0",
    "plotly>=6.0.1",
    "prettytable>=3.16.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "scikit-learn>=1.6.1",
    "streamlit>=1.45.1",
]
